+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed aborting Snapshot procedure (make sure resume-io is executed)
 - Fixed AutoSnapshot scheduling bug
 - Fixed rollback after clearing entire PropsContainers
+- Fixed that too many concurrent Snapshots led to unnecessarily long suspend-io state
 
 ## [1.20.3] - 2023-01-26